+++ src/App.js
@@ -1,11 +1,27 @@
 import React from "react";
-import Events from './pages/my-events'
+import Wrapper from "./components/Wrapper/Wrapper"
+import Events from './pages/MyEvents'
+import Home from './pages/Home'
+import { BrowserRouter as Router, Route } from "react-router-dom";
 
 
 function App() {
-    return <Events />
-    
+    return( 
+        <Router>
+               <div>
+         <Wrapper>
+             <Route exact path ="/" component={Home}/>
+             <Route exact path ="/events" component={Events}/>
 
+
+            </Wrapper>
+            </div>
+
+        </Router>
+    
+   
+    
+    )
 }
 
 export default App;
