+++ src/components/Wrapper/Wrapper.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+
+
+function Wrapper(props) {
+  return <main className="wrapper">{props.children}</main>;
+}
+
+export default Wrapper;
