+++ src/pages/Home.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+function Home(){
+
+    return(
+        <p> Hello World</p>
+    )
+
+}
+
+export default Home;
